Provided ability to remove rate.

+++ Script_odont/Core/SctFinished.swift
@@ -20,5 +20,5 @@ struct SctFinished
     var statistics: SctStatistics
     
     var score: Double
-    var vote: Double?
+    var vote: Int?
 }
